requirements.txt
porntrex_api
aiohttp
Pillow